requirements.txt
Django==3.2.18
PyJWT==2.5.0
boto3==1.24.77
cryptography==3.4.8
django-cors-headers==3.13.0
django-filter==22.1
https://github.com/LyridInc/django-graphql-auth/tarball/master
django-model-utils==4.2.0
django-phonenumber-field==7.0.0
django-ses==3.1.2
django-templated-email==3.0.0
django-timezone-field==5.0
dnspython==1.16.0
graphene-django==2.15.0
https://github.com/LyridInc/graphene-django-cud/tarball/master
phonenumbers==8.12.55
python-dotenv==0.21.0
requests==2.28.1
social-auth-core[all]==4.4.0
social-auth-app-django==5.1.0
whitenoise==6.2.0
sentry-sdk==1.10.1

psycopg2-binary==2.9.3
graphene==2.1.9
dj-database-url==1.0.0

pytest>=6.2.5
pytest-cov>=3.0
pytest-django==4.5.2
mixer==7.2.2
tzdata==2022.7